requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
telegram
python-telegram-bot==21.0